// apps/admin/src/app/types/docs.types.ts
export type CodeLanguage = 'typescript' | 'html' | 'scss' | 'bash' | 'json';

/**
 * Code tab configuration
 */
export interface CodeTab {
  label: string;
  code: string;
  language: CodeLanguage;
}